requirements.txt
torch
torchvision
pandas
gradio
Pillow